+++ R/dfm_lookup.R
@@ -80,6 +80,7 @@ dfm_lookup.dfm <- function(x, dictionary, levels = 1:5,
     valuetype <- match.arg(valuetype)
     attrs <- attributes(x)
     lengths <- ntoken(x)
+    types <- colnames(x)
     
     # Generate all combinations of type IDs
     values_id <- c()
@@ -89,22 +90,23 @@ dfm_lookup.dfm <- function(x, dictionary, levels = 1:5,
         catm("applying a dictionary consisting of ", length(dictionary), " key", 
              if (length(dictionary) > 1L) "s" else "", "\n", sep="")
     
+    index <- index_types(types, valuetype, case_insensitive) # index types before the loop
     for (h in seq_along(dictionary)) {
         values <- as.list(stri_replace_all_fixed(dictionary[[h]], ' ', attr(x, 'concatenator')))
-        values_temp <- unlist(regex2id(values, colnames(x), valuetype, case_insensitive, FALSE))
+        values_temp <- unlist(regex2id(values, index = index))
         values_id <- c(values_id, values_temp)
         keys_id <- c(keys_id, rep(h, length(values_temp)))
     }
-    
     if (length(values_id)) {
         keys <- names(dictionary)
         if (capkeys)
             keys <- char_toupper(keys)
         if (exclusive) {
             if (!is.null(nomatch)) {
-                values_id <- c(values_id, setdiff(seq_len(nfeature(x)), values_id))
-                keys_id <- c(keys_id, rep(length(keys) + 1, nfeature(x) - length(keys_id)))
-                keys <- c(keys, nomatch)
+                values_id_nomatch <- setdiff(seq_len(nfeature(x)), values_id)
+                values_id <- c(values_id, values_id_nomatch)
+                keys_id <- c(keys_id, rep(max(keys_id) + 1, length(values_id_nomatch)))
+                keys <- c(keys, nomatch[1])
             }
             x <- x[,values_id]
             cols_new <- keys[keys_id]
@@ -115,7 +117,7 @@ dfm_lookup.dfm <- function(x, dictionary, levels = 1:5,
         } else {
             if (!is.null(nomatch))
                 warning("nomatch only applies if exclusive = TRUE")
-            cols_new <- colnames(x)
+            cols_new <- types
             cols_new[values_id] <- keys[keys_id]
             colnames(x) <- cols_new
             result <- dfm_compress(x, margin = 'features')