+++ R/tokens.R
@@ -226,7 +226,7 @@ tokens.tokens <-  function(x, what = c("word", "sentence", "character", "fastest
                            concatenator = "_",
                            hash = TRUE,
                            verbose = quanteda_options("verbose"),
-                           include_docvars,
+                           include_docvars = TRUE,
                            ...) {
     
     types <- types(x)
@@ -252,10 +252,12 @@ tokens.tokens <-  function(x, what = c("word", "sentence", "character", "fastest
     
     if (length(regex))
         x <- tokens_remove(x, paste(regex, collapse = '|'), valuetype = 'regex', padding = FALSE)
-    if (!identical(ngrams, 1L))
+    if (!identical(ngrams, 1L) || !identical(skip, 0L))
         x <- tokens_ngrams(x, n = ngrams, skip = skip, concatenator = concatenator)
     if (hash == FALSE)
         x <- as.tokenizedTexts(x)
+    if (!include_docvars)
+        docvars(x) <- data.frame(row.names = docnames(x))
     return(x)
 }