+++ R/utils.R
@@ -223,28 +223,54 @@ code <- function(texts){
 }
 
 
-# convert various input as features to a vector used in tokens_select, 
-# tokens_compound and kwic.
-features2id <- function(features, types, valuetype, case_insensitive, 
-                        concatenator = '_', remove_unigram = FALSE) {
+#' convert various input as pattern to a vector used in tokens_select, 
+#' tokens_compound and kwic.
+#' @inheritParams pattern
+#' @inheritParams valuetype
+#' @param case_insensitive ignore the case of dictionary values if \code{TRUE}
+#' @param concatenator concatenator that join multi-word expression in tokens object
+#' @param remove_unigram ignore single-word patterns if \code{TRUE}
+#' @seealso regex2id
+#' @keywords internal
+pattern2id <- function(pattern, types, valuetype, case_insensitive, 
+                       concatenator = '_', remove_unigram = FALSE) {
     
-    if (is.sequences(features) || is.collocations(features)) {
-        features <- stri_split_charclass(features$collocation, "\\p{Z}")
-        features_id <- lapply(features, function(x) fastmatch::fmatch(x, types))
-        features_id <- features_id[sapply(features_id, function(x) all(!is.na(x)))]
+    if (is.sequences(pattern) || is.collocations(pattern)) {
+        pattern <- stri_split_charclass(pattern$collocation, "\\p{Z}")
+        pattern_id <- lapply(pattern, function(x) fastmatch::fmatch(x, types))
+        pattern_id <- pattern_id[sapply(pattern_id, function(x) all(!is.na(x)))]
     } else {
-        if (is.dictionary(features)) {
-            features <- unlist(features, use.names = FALSE)
-            features <- split_dictionary_values(features, concatenator)
+        if (is.dictionary(pattern)) {
+            pattern <- unlist(pattern, use.names = FALSE)
+            pattern <- split_dictionary_values(pattern, concatenator)
         } else {
-            features <- as.list(features)
+            pattern <- as.list(pattern)
         }
         if (remove_unigram)
-            features <- features[lengths(features) > 1] # drop single-word features
-        features_id <- regex2id(features, types, valuetype, case_insensitive)
+            pattern <- pattern[lengths(pattern) > 1] # drop single-word pattern
+        pattern_id <- regex2id(pattern, types, valuetype, case_insensitive)
     }
-    attr(features_id, 'features') <- stri_c_list(features, sep = ' ')
-    return(features_id)
+    attr(pattern_id, 'pattern') <- stri_c_list(pattern, sep = ' ')
+    return(pattern_id)
+}
+
+
+#' internal function for \code{select_types()} to check if a string is a regular expression
+#' @param x a character string to be tested
+#' @keywords internal
+is_regex <- function(x){
+    any(stri_detect_fixed(x, c(".", "(", ")", "^", "{", "}", "+", "$", "*", "?", "[", "]", "\\")))
+}
+
+#' internal function for \code{select_types()} to escape regular expressions 
+#' 
+#' This function escapes glob patterns before \code{utils:glob2rx()}, therefore * and ?
+#' are unescaped.
+#' @param x character vector to be escaped
+#' @keywords internal
+escape_regex <- function(x){
+    #stri_replace_all_regex(x, "([.()^\\{\\}+$*\\[\\]\\\\])", "\\\\$1") # escape any
+    stri_replace_all_regex(x, "([.()^\\{\\}+$\\[\\]\\\\])", "\\\\$1") # allow glob
 }
 
 # which object class started the call stack?
@@ -259,5 +285,16 @@ who_called_me_first <- function(x, function_name) {
     x <- x[base_call_index[-1]]
     x <- stringi::stri_replace_all_regex(x, paste0(function_name, "\\.(\\w+)\\(.+$"), "$1")
     x[1]
+}
+
+# function to check dots arguments against a list of permissible arguments
+check_dots <-  function(dots, permissible_args = NULL) {
+    if (length(dots) == 0) return()
+    args <- names(dots)
+    impermissible_args <-  setdiff(args, permissible_args)
+    if (length(impermissible_args))
+        warning("Argument", if (length(impermissible_args) > 1) "s " else " ", 
+                paste(impermissible_args, collapse = ', '), " not used.", 
+                noBreaks. = TRUE, call. = FALSE)
 }